+++ src/common/universalelectricity/electricity/ElectricityTransferData.java
@@ -6,11 +6,11 @@ import universalelectricity.extend.IElectricUnit;
 public class ElectricityTransferData
 {
     public IElectricUnit electricUnit;
-    public float amps;
-    public float voltage;
+    public double amps;
+    public double voltage;
     public ForgeDirection side;
 
-    public ElectricityTransferData(IElectricUnit electricUnit, ForgeDirection side, float amps, float voltage)
+    public ElectricityTransferData(IElectricUnit electricUnit, ForgeDirection side, double amps, double voltage)
     {
         this.electricUnit = electricUnit;
         this.side = side;
